fix:clean arc in handle data in fe and minor bug issues in get data profile
