+++ frontend/src/app/layout.tsx
@@ -14,7 +14,7 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "E-commerce system",
+  title: "NOIS",
   description: "Generated by create next app",
 };
